refactor namespace package - cleanup opencensus root folder
