+++ opencensus/trace/exporters/ocagent/trace_exporter.py
@@ -19,7 +19,7 @@ import grpc
 import os
 import socket
 
-from opencensus.__version__ import __version__
+from opencensus.common.__version__ import __version__
 from opencensus.common.transports import sync
 from opencensus.trace.exporters import base
 from opencensus.trace.exporters.gen.opencensus.agent.common.v1 \
